rename: 예약을 찾을 수 없을 때 발생하는 예외의 패키지 위치 변경
