+++ src/main/java/roomescape/controller/exception/ReservationNotFoundException.java
@@ -1,4 +1,4 @@
-package roomescape.exception;
+package roomescape.controller.exception;
 
 public class ReservationNotFoundException extends IllegalArgumentException {
